Added HTML Templates for Landing Page (index), Dashboard, and Edit Quote
In this commit, I have introduced the initial HTML templates for three crucial parts of the application: the landing page (index.html), the user dashboard
(dashboard.html), and the edit quote page (edit_quote.html). These templates extend the base.html template, setting the stage for the web pages in the
project.

- index.html: It sets up the landing page, complete with links for users to access the login and signup sections.
- dashboard.html: This template establishes the primary hub for logged-in users, offering spaces for displaying and editing quotes.
- edit_quote.html: It provides a dedicated page for users to modify the content of their quotes through a form.

These templates will serve as the foundation of the user interface. Routes will be defined further along build.